requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-rest-auth==2.1.11
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
